+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 LifterLMS Changelog
 ===================
 
+v3.17.3 - 2018-04-10
+--------------------
+
++ Fix issue causing current date / time to display as the End Date for incomplete quiz attempts on quiz reporting screens
+
+
 v3.17.2 - 2018-04-09
 --------------------